Projects/MessageQueueExperiments/kafka_partitioning/requirements.txt
kafka-python==2.0.2
python-dotenv==1.0.0
colorama==0.4.6
